# src/configs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Blog",
    "description": "Ejemplo de API para manejo de un Blog.",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "User actions"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/v1/users": {
      "post": {
        "tags": ["Users"],
        "description": "Create user",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "model",
            "required": ["user_id", "username", "password"],
            "in": "body",
            "properties": {
              "profile_id": {
                "$ref": "#/definitions/user_id"
              },
              "username": {
                "$ref": "#/definitions/username"
              },
              "password": {
                "$ref": "#/definitions/password"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/server_error"
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "description": "Update user",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "model",
            "required": ["user_id", "username", "password"],
            "in": "body",
            "properties": {
              "user_id": {
                "$ref": "#/definitions/user_id"
              },
              "username": {
                "$ref": "#/definitions/username"
              },
              "password": {
                "$ref": "#/definitions/password"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/server_error"
            }
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "description": "List users",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "$ref": "#/definitions/user_id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/server_error"
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "description": "Delete user",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "$ref": "#/definitions/user_id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/server_error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "user_id": {
      "type": "number",
      "description": "Id for User",
      "example": 1
    },
    "username": {
      "type": "string",
      "description": "Username for login",
      "example": "luke"
    },
    "password": {
      "type": "string",
      "description": "User password for login",
      "example": "1234"
    },
    "server_error": {
      "type": "object",
      "description": "Server Error",
      "properties": {
        "msg": {
          "type": "string",
          "description": "Server Error",
          "example": "Server Error"
        }
      }
    }
  }
}